+++ packages/expo-router/src/react-navigation/routers/TabRouter.tsx
@@ -1,6 +1,7 @@
 import { nanoid } from 'nanoid/non-secure';
 
 import { BaseRouter } from './BaseRouter';
+import { areRouteNamesEqual } from './areRouteNamesEqual';
 import { createParamsFromAction } from './createParamsFromAction';
 import type {
   CommonNavigationAction,
@@ -255,7 +256,7 @@ export function TabRouter({
     getRehydratedState(partialState, { routeNames, routeParamList }) {
       const state = partialState;
 
-      if (state.stale === false) {
+      if (state.stale === false && areRouteNamesEqual(state.routeNames, routeNames)) {
         return state;
       }
 
@@ -305,36 +306,6 @@ export function TabRouter({
         backBehavior,
         initialRouteName
       );
-    },
-
-    getStateForRouteNamesChange(state, { routeNames, routeParamList, routeKeyChanges }) {
-      const routes = routeNames.map(
-        (name) =>
-          state.routes.find((r) => r.name === name && !routeKeyChanges.includes(r.name)) || {
-            name,
-            key: `${name}-${nanoid()}`,
-            params: routeParamList[name],
-          }
-      );
-
-      const index = Math.max(0, routeNames.indexOf(state.routes[state.index]!.name));
-
-      let history = state.history.filter(
-        // Type will always be 'route' for tabs, but could be different in a router extending this (e.g. drawer)
-        (it) => it.type !== 'route' || routes.find((r) => r.key === it.key)
-      );
-
-      if (!history.length) {
-        history = getRouteHistory(routes, index, backBehavior, initialRouteName);
-      }
-
-      return {
-        ...state,
-        history,
-        routeNames,
-        routes,
-        index,
-      };
     },
 
     getStateForRouteFocus(state, key) {