+++ packages/expo-router/CHANGELOG.md
@@ -44,6 +44,7 @@
 
 ### 💡 Others
 
+- Remove render time set state on `routeName` changes ([#47986](https://github.com/expo/expo/pull/47986) by [@Ubax](https://github.com/Ubax))
 - Refactor headless tabs not to rely on `routeNames` ([#48257](https://github.com/expo/expo/pull/48257) by [@Ubax](https://github.com/Ubax))
 - Add support for replace action in tabs ([#48256](https://github.com/expo/expo/pull/48256) by [@Ubax](https://github.com/Ubax))
 - Integrate native stack with standard navigation ([#48114](https://github.com/expo/expo/pull/48114) by [@Ubax](https://github.com/Ubax))